Demo in separate activity and input spoofing in lower split
