+++ server/src/main/AndroidManifest.xml
@@ -16,6 +16,15 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".EventGrabDemo"
+            android:label="@string/title_activity_event_grab_demo"
+            android:parentActivityName=".MainActivity"
+            android:theme="@style/AppTheme.NoActionBar">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value="com.amos.server.MainActivity" />
+        </activity>
     </application>
 
 </manifest>
